% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{am_argreduce}
\alias{am_argreduce}
\alias{am_rowargmax}
\alias{am_rowargmin}
\alias{am_colargmax}
\alias{am_colargmin}
\title{Row and column argmax/argmin}
\usage{
am_rowargmax(x)

am_rowargmin(x)

am_colargmax(x)

am_colargmin(x)
}
\arguments{
\item{x}{A numeric matrix or \code{adgeMatrix}.}
}
\value{
An integer vector of indices.
}
\description{
Return the index of the maximum or minimum value in each row or column.
}
