% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-aliases.R
\name{amatrix-methods-internal}
\alias{amatrix-methods-internal}
\alias{[,adgCMatrix,ANY,ANY,ANY-method}
\alias{[,adgCMatrix,index,index,logical-method}
\alias{[,adgCMatrix,index,missing,logical-method}
\alias{[,adgCMatrix,missing,index,logical-method}
\alias{[,adgeMatrix,ANY,ANY,ANY-method}
\alias{[,adgeMatrix,index,index,logical-method}
\alias{[,adgeMatrix,index,missing,logical-method}
\alias{[,adgeMatrix,missing,index,logical-method}
\alias{[<-,adgCMatrix,ANY,ANY,ANY-method}
\alias{[<-,adgCMatrix,index,index,Matrix-method}
\alias{[<-,adgCMatrix,index,index,integer-method}
\alias{[<-,adgCMatrix,index,index,logical-method}
\alias{[<-,adgCMatrix,index,index,matrix-method}
\alias{[<-,adgCMatrix,index,index,numeric-method}
\alias{[<-,adgeMatrix,ANY,ANY,ANY-method}
\alias{[<-,adgeMatrix,index,index,Matrix-method}
\alias{[<-,adgeMatrix,index,index,integer-method}
\alias{[<-,adgeMatrix,index,index,logical-method}
\alias{[<-,adgeMatrix,index,index,matrix-method}
\alias{[<-,adgeMatrix,index,index,numeric-method}
\alias{crossprod,KronMatrix,matrix-method}
\alias{crossprod,KronMatrix,missing-method}
\alias{crossprod,KronMatrix,numeric-method}
\alias{crossprod,dgCMatrix,adgCMatrix-method}
\alias{crossprod,dgeMatrix,adgCMatrix-method}
\alias{crossprod,matrix,adgCMatrix-method}
\alias{crossprod,matrix,adgeMatrix-method}
\alias{crossprod,numeric,adgCMatrix-method}
\alias{crossprod,numeric,adgeMatrix-method}
\alias{diag,adgCMatrix-method}
\alias{diag,adgeMatrix-method}
\alias{dim,KronMatrix-method}
\alias{dim,aTransposeView-method}
\alias{dimnames,aTransposeView-method}
\alias{dimnames<-,adgCMatrix,ANY-method}
\alias{dimnames<-,adgeMatrix,ANY-method}
\alias{eigen,matrix-method}
\alias{norm,adgCMatrix,ANY-method}
\alias{norm,adgeMatrix,ANY-method}
\alias{norm,matrix,ANY-method}
\alias{norm,numeric,ANY-method}
\alias{qr.Q,amQR-method}
\alias{qr.R,amQR-method}
\alias{qr.coef,amQR,ANY-method}
\alias{qr.fitted,amQR,ANY-method}
\alias{qr.qty,amQR,ANY-method}
\alias{qr.qy,amQR,ANY-method}
\alias{qr.resid,amQR,ANY-method}
\alias{qr.solve,amQR,ANY-method}
\alias{qr.solve,amQR,missing-method}
\alias{qr,adgCMatrix-method}
\alias{qr,adgeMatrix-method}
\alias{solve,KronMatrix,matrix-method}
\alias{solve,KronMatrix,missing-method}
\alias{solve,KronMatrix,numeric-method}
\alias{svd,matrix-method}
\alias{t,KronMatrix-method}
\alias{t,aTransposeView-method}
\alias{t,adgCMatrix-method}
\alias{t,adgeMatrix-method}
\alias{tcrossprod,dgCMatrix,adgCMatrix-method}
\alias{tcrossprod,dgeMatrix,adgCMatrix-method}
\alias{tcrossprod,matrix,adgCMatrix-method}
\alias{tcrossprod,matrix,adgeMatrix-method}
\alias{tcrossprod,numeric,adgCMatrix-method}
\title{Internal S4 method dispatch entries}
\description{
These S4 methods are dispatch implementations of standard generics
(`%*%`, `[`, `crossprod`, `t`, etc.) for amatrix classes. They are
not part of the public API -- use the generics directly. This help
page exists only to satisfy R CMD check.
}
\keyword{internal}
