% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amatrix-package.R
\docType{package}
\name{amatrix-package}
\alias{amatrix}
\alias{amatrix-package}
\title{amatrix package}
\description{
Package-level roxygen directives for namespace generation.
}
\author{
\strong{Maintainer}: Brad Buchsbaum \email{brad.buchsbaum@gmail.com}

}
\keyword{internal}
