% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{batch_crossprod}
\alias{batch_crossprod}
\title{Batch crossproduct}
\usage{
batch_crossprod(A)
}
\arguments{
\item{A}{A list of numeric matrices, or a 3-D array \code{[n, p, B]}.}
}
\value{
A list of \code{p x p} crossproduct matrices.
}
\description{
Compute \code{t(A_b) \%*\% A_b} for each matrix in a batch.
}
\seealso{
\code{\link{batch_chol}}
}
