% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{batch_solve}
\alias{batch_solve}
\title{Batch triangular solve}
\usage{
batch_solve(Ls, B)
}
\arguments{
\item{Ls}{A list of \code{amChol} objects (output of \code{batch_chol}).}

\item{B}{A list of right-hand-side matrices/vectors, or a 3-D array
\code{[n, k, B]}.  Length / third dimension must match \code{Ls}.}
}
\value{
A list of solution matrices (or vectors when each rhs is a vector).
}
\description{
Solve B linear systems \code{A_b x_b = B_b} where each \code{A_b} is
represented by its Cholesky factor from \code{\link{batch_chol}}.
}
\seealso{
\code{\link{batch_chol}}, \code{\link{chol_solve}}
}
