% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chol-factor.R
\name{chol_diag}
\alias{chol_diag}
\title{Extract the diagonal of a Cholesky factor}
\usage{
chol_diag(factor)
}
\arguments{
\item{factor}{An \code{\linkS4class{amChol}} object from
\code{\link{chol_factor}}.}
}
\value{
Numeric vector of length equal to the matrix dimension.
}
\description{
Returns the diagonal of the upper-triangular matrix \code{R} stored
in an \code{\linkS4class{amChol}} object.
}
\examples{
m <- crossprod(matrix(rnorm(16), 4, 4)) + diag(4)
A <- adgeMatrix(m)
fac <- chol_factor(A)
chol_diag(fac)

}
