% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chol-factor.R
\name{chol_factor}
\alias{chol_factor}
\title{Compute the Cholesky factorization of an adgeMatrix}
\usage{
chol_factor(X)
}
\arguments{
\item{X}{An \code{adgeMatrix} that is symmetric positive definite.}
}
\value{
An \code{\linkS4class{amChol}} object.
}
\description{
Computes the upper-triangular Cholesky factor \code{R} of a
symmetric positive-definite \code{adgeMatrix} \code{X} such that
\code{t(R) \%*\% R == X}. Results are cached by \code{object_id};
repeated calls with the same object return the cached factor.
}
\examples{
m <- crossprod(matrix(rnorm(16), 4, 4)) + diag(4)
A <- adgeMatrix(m)
fac <- chol_factor(A)
fac

}
