% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-coercion.R
\name{as.matrix,adgeMatrix-method}
\alias{as.matrix,adgeMatrix-method}
\alias{as.matrix,adgCMatrix-method}
\alias{as.matrix,aTransposeView-method}
\alias{as.numeric,adgeMatrix-method}
\alias{as.vector,adgeMatrix-method}
\alias{as.array,adgeMatrix-method}
\alias{as.array,adgCMatrix-method}
\title{Coerce amatrix objects to base R types}
\usage{
\S4method{as.matrix}{adgeMatrix}(x, ...)

\S4method{as.matrix}{adgCMatrix}(x, ...)

\S4method{as.matrix}{aTransposeView}(x, ...)

\S4method{as.numeric}{adgeMatrix}(x, ...)

\S4method{as.vector}{adgeMatrix}(x, mode = "any")

\S4method{as.array}{adgeMatrix}(x, ...)

\S4method{as.array}{adgCMatrix}(x, ...)
}
\arguments{
\item{x}{An \code{adgeMatrix}, \code{adgCMatrix}, or
\code{aTransposeView}.}

\item{...}{Further arguments passed to the corresponding base R
coercion function.}

\item{mode}{Storage mode string passed to \code{as.vector}.}
}
\value{
A plain R \code{matrix}, numeric vector, or \code{array}
  containing the materialized host data.
}
\description{
Convert \code{adgeMatrix}, \code{adgCMatrix}, or \code{aTransposeView}
objects to base R \code{matrix}, numeric vector, or array by
materializing the host copy.
}
\examples{
A <- adgeMatrix(matrix(1:6, 2, 3))
as.matrix(A)

}
