% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sparse.R
\name{crossprod,adgCMatrix,missing-method}
\alias{crossprod,adgCMatrix,missing-method}
\alias{crossprod,adgCMatrix,ANY-method}
\alias{crossprod,adgCMatrix,matrix-method}
\alias{crossprod,adgCMatrix,Matrix-method}
\alias{crossprod,adgCMatrix,dgeMatrix-method}
\alias{crossprod,adgCMatrix,dgCMatrix-method}
\alias{crossprod,adgCMatrix,adgeMatrix-method}
\alias{crossprod,adgCMatrix,adgCMatrix-method}
\alias{tcrossprod,adgCMatrix,missing-method}
\alias{tcrossprod,adgCMatrix,ANY-method}
\alias{tcrossprod,adgCMatrix,matrix-method}
\alias{tcrossprod,adgCMatrix,Matrix-method}
\alias{tcrossprod,adgCMatrix,dgeMatrix-method}
\alias{tcrossprod,adgCMatrix,dgCMatrix-method}
\alias{tcrossprod,adgCMatrix,adgeMatrix-method}
\alias{tcrossprod,adgCMatrix,adgCMatrix-method}
\title{Cross-product methods for adgCMatrix}
\usage{
\S4method{crossprod}{adgCMatrix,missing}(x, y = NULL, ...)

\S4method{crossprod}{adgCMatrix,ANY}(x, y = NULL, ...)

\S4method{crossprod}{adgCMatrix,matrix}(x, y = NULL, ...)

\S4method{crossprod}{adgCMatrix,Matrix}(x, y = NULL, ...)

\S4method{crossprod}{adgCMatrix,dgeMatrix}(x, y = NULL, ...)

\S4method{crossprod}{adgCMatrix,dgCMatrix}(x, y = NULL, ...)

\S4method{crossprod}{adgCMatrix,adgeMatrix}(x, y = NULL, ...)

\S4method{crossprod}{adgCMatrix,adgCMatrix}(x, y = NULL, ...)

\S4method{tcrossprod}{adgCMatrix,missing}(x, y = NULL, ...)

\S4method{tcrossprod}{adgCMatrix,ANY}(x, y = NULL, ...)

\S4method{tcrossprod}{adgCMatrix,matrix}(x, y = NULL, ...)

\S4method{tcrossprod}{adgCMatrix,Matrix}(x, y = NULL, ...)

\S4method{tcrossprod}{adgCMatrix,dgeMatrix}(x, y = NULL, ...)

\S4method{tcrossprod}{adgCMatrix,dgCMatrix}(x, y = NULL, ...)

\S4method{tcrossprod}{adgCMatrix,adgeMatrix}(x, y = NULL, ...)

\S4method{tcrossprod}{adgCMatrix,adgCMatrix}(x, y = NULL, ...)
}
\arguments{
\item{x}{An \code{adgCMatrix}.}

\item{y}{A matrix-like object or \code{NULL}/missing for the symmetric
form.}

\item{...}{Further arguments passed to the backend.}
}
\value{
An \code{adgeMatrix} or \code{adgCMatrix} containing the result.
}
\description{
Compute \eqn{t(x) \%*\% y} (\code{crossprod}) or
\eqn{x \%*\% t(y)} (\code{tcrossprod}) for sparse \code{adgCMatrix}
objects, dispatching through the amatrix backend.
}
\examples{
sp <- as(matrix(c(1, 0, 0, 2, 1, 0), 3, 2), "dgCMatrix")
A  <- adgCMatrix(sp)
crossprod(A)

}
