% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_adlgCMatrix}
\alias{new_adlgCMatrix}
\title{Construct an adlgCMatrix from a sparse or dense logical matrix}
\usage{
new_adlgCMatrix(
  x,
  preferred_backend = .amatrix_default_preferred_backend(policy, precision),
  policy = amatrix_default_policy(),
  precision = amatrix_default_precision()
)
}
\description{
Construct an adlgCMatrix from a sparse or dense logical matrix
}
\keyword{internal}
