% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_adlgeMatrix}
\alias{new_adlgeMatrix}
\title{Construct an adlgeMatrix from a matrix or lgeMatrix}
\usage{
new_adlgeMatrix(
  x,
  preferred_backend = .amatrix_default_preferred_backend(policy, precision),
  policy = amatrix_default_policy(),
  precision = amatrix_default_precision(),
  src_id = ""
)
}
\description{
Construct an adlgeMatrix from a matrix or lgeMatrix
}
\keyword{internal}
