% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{trace}
\alias{trace}
\title{Matrix trace}
\usage{
trace(x)
}
\arguments{
\item{x}{A square matrix, sparse \code{sparseMatrix}, or \code{aMatrix}.}
}
\value{
A single numeric scalar equal to the sum of diagonal elements.
}
\description{
Returns the trace (sum of diagonal elements) of a square matrix or
\code{aMatrix}.
}
\examples{
trace(diag(1:4))

}
