% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amatrix-package.R
\docType{package}
\name{amatrix-package}
\alias{amatrix}
\alias{amatrix-package}
\title{amatrix package}
\description{
Package-level roxygen directives for namespace generation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bbuchsbaum.github.io/amatrix/}
  \item \url{https://github.com/bbuchsbaum/amatrix}
  \item Report bugs at \url{https://github.com/bbuchsbaum/amatrix/issues}
}

}
\author{
\strong{Maintainer}: Brad Buchsbaum \email{brad.buchsbaum@gmail.com}

}
\keyword{internal}
